Output the estimated distances' variances alongside the distance matrix
PhyloTree::computeDist(() now writes a separate file containing the
variances. Note that whether or not the variances are computed depends
on the various parameters, currently it seems that "-lsvar" has to be
set to "second_taylor" (NOT the default), that Newton's method (and not
Brent's) must be used, and possibly a few other things.